+++ advanced_tutorials/nyc_taxi_fares/README.md
@@ -45,3 +45,4 @@ or
 ![1.png](streamlit_examples/1.png)
 ![2.png](streamlit_examples/2.png)
 ![3.png](streamlit_examples/3.png)
+![4.png](streamlit_examples/4.png)
