+++ advanced_tutorials/nyc_taxi_fares/temp_coordinates.json
@@ -1,1 +1,1 @@
-{"c1": {"lat": 40.84706035607122, "long": -72.77893066406251, "time_clicked": 1664916244.434076}, "c2": {"lat": 40.84706035607122, "long": -72.77893066406251, "time_clicked": 1664916167.9903588}}
+{"c1": {"lat": 40.68063802521456, "long": -73.38317871093751, "time_clicked": 1665761827.854382}, "c2": {"lat": 40.68063802521456, "long": -73.38317871093751, "time_clicked": 1665761732.2754767}}
